requirements.txt
requests>=2.31
python-dotenv>=1.0
groq>=0.11.0
